AND product (java8) [20/20] Accepted

+++ master/and-product.java
@@ -6,10 +6,10 @@ public class Solution {
     public static void main(String[] args) {
         Scanner in = new Scanner(System.in);
         in.nextInt();
-        while(in.hasNextInt()) {
-            int x = in.nextInt();
-            int y = in.nextInt();
-            int res = x;
+        while(in.hasNextLong()) {
+            long x = in.nextLong();
+            long y = in.nextLong();
+            long res = x;
             while(x++ < y)
                 res &= x;
             System.out.println(res);
